Move development harness to harness.py
